Registers: Store a bounded history of yanks
This doesn't fundamentally change `Registers` - it extends the
underlying storage of each regular `Register` (a sequence of `String`)
to support a history of yanked values by storing a limited set of past
yanks and tracking the length of each yank.

This commit also changes `Registers::write` to accept an IntoIterator of
the values. But otherwise this compiles with no changes necessary to
callers and no behavior changes visible from outside the module.

History is limited at 10 entries to set a rough bounds on memory usage.